requirements.txt
pandas>=1.3.0
pyyaml>=6.0